added dev mode, leaving more packages for interactive use

+++ script/minimize.sh
@@ -11,14 +11,17 @@ echo "==> Removing all linux kernels except the currrent one"
 dpkg --list | awk '{ print $2 }' | grep 'linux-image-3.*-generic' | grep -v $(uname -r) | xargs apt-get -y purge
 echo "==> Removing linux headers"
 dpkg --list | awk '{ print $2 }' | grep linux-headers | xargs apt-get -y purge
-rm -rf /usr/src/linux-headers*
 echo "==> Removing linux source"
 dpkg --list | awk '{ print $2 }' | grep linux-source | xargs apt-get -y purge
-echo "==> Removing development packages"
-dpkg --list | awk '{ print $2 }' | grep -- '-dev$' | xargs apt-get -y purge
-echo "==> Removing documentation"
-dpkg --list | awk '{ print $2 }' | grep -- '-doc$' | xargs apt-get -y purge
-apt-get -y purge build-essential
+
+if ! [[ ${DEV} == true ]]; then
+    echo "==> Removing development packages"
+    dpkg --list | awk '{ print $2 }' | grep -- '-dev$' | xargs apt-get -y purge
+    echo "==> Removing documentation"
+    dpkg --list | awk '{ print $2 }' | grep -- '-doc$' | xargs apt-get -y purge
+    apt-get -y purge build-essential
+fi
+
 echo "==> Removing X11 libraries"
 apt-get -y purge libx11-data xauth libxmuu1 libxcb1 libx11-6 libxext6
 echo "==> Removing obsolete networking components"
@@ -49,16 +52,21 @@ while [ -n "$(deborphan --guess-all --libdevel)" ]; do
 done
 apt-get -y purge deborphan dialog
 
-echo "==> Removing man pages"
-find /usr/share/man -type f -delete
+if ! [[ ${DEV} == true ]]; then
+    echo "==> Removing man pages"
+    find /usr/share/man -type f -delete
+    echo "==> Removing any docs"
+    find /usr/share/doc -type f -delete
+
+    echo "==> Removing groff info lintian linda"
+    rm -rf /usr/share/groff/* /usr/share/info/* /usr/share/lintian/* /usr/share/linda/*
+fi
+
 echo "==> Removing APT files"
 find /var/lib/apt -type f -delete
-echo "==> Removing any docs"
-find /usr/share/doc -type f -delete
 echo "==> Removing caches"
 find /var/cache -type f -delete
-echo "==> Removing groff info lintian linda"
-rm -rf /usr/share/groff/* /usr/share/info/* /usr/share/lintian/* /usr/share/linda/*
+echo "==> Removing lintian linda"
 
 echo "==> Disk usage after cleanup"
 df -h
